maipoc-main/maipoc-main/backend/requirements.txt
# aiohttp==3.8.1
# aiosignal==1.2.0
# alembic==1.8.0
# argon2-cffi==21.3.0
# argon2-cffi-bindings==21.2.0
# async-timeout==4.0.2
# asynctest==0.13.0
# attrs==21.4.0
# azure-core==1.24.1
# azure-storage-blob==12.12.0
# azure-storage-file-datalake==12.7.0
# backcall==0.2.0
# beautifulsoup4==4.11.1
# bleach==5.0.0
# blis==0.7.7
# Boruta==0.3
# catalogue==1.0.0
# category-encoders==2.4.0
# certifi==2022.6.15
# cffi==1.15.0
# charset-normalizer==2.0.12
# click==8.1.3
# cloudpickle==2.1.0
# colorama==0.4.4
# colorlover==0.3.0
# cryptography==37.0.2
# cufflinks==0.17.3
# cycler==0.11.0
# cymem==2.0.6
# Cython==0.29.28
# databricks-cli==0.16.8
# debugpy==1.6.0
# decorator==5.1.1
# defusedxml==0.7.1
# docker==5.0.3
# entrypoints==0.4
# fastjsonschema==2.15.3
# Flask==2.1.2
# fonttools==4.33.3
# frozenlist==1.3.0
# funcy==1.17
# future==0.18.2
# gensim==4.2.0
# gitdb==4.0.9
# GitPython==3.1.27
# gower==0.0.5
# greenlet==1.1.2
# htmlmin==0.1.12
# idna==3.3
# ImageHash==4.2.1
# imbalanced-learn==0.7.0
# importlib-metadata==4.11.4
# importlib-resources==5.8.0
# ipykernel==6.15.0
# ipython==7.34.0
# ipython-genutils==0.2.0
# ipywidgets==7.7.0
# isodate==0.6.1
# itsdangerous==2.1.2
# jedi==0.18.1
# Jinja2==3.1.2
# joblib==1.1.0
# jsonschema==4.6.0
# jupyter-client==7.3.4
# jupyter-core==4.10.0
# jupyterlab-pygments==0.2.2
# jupyterlab-widgets==1.1.0
# kiwisolver==1.4.3
# kmodes==0.12.1
# lightgbm==3.3.2
# llvmlite==0.38.1
# Mako==1.2.0
# MarkupSafe==2.1.1
# matplotlib==3.5.2
# matplotlib-inline==0.1.3
# missingno==0.5.1
# mistune==0.8.4
# mlflow==1.26.1
# mlxtend==0.19.0
# msrest==0.7.1
# multidict==6.0.2
# multimethod==1.8
# murmurhash==1.0.7
# nbclient==0.6.4
# nbconvert==6.5.0
# nbformat==5.4.0
# nest-asyncio==1.5.5
# networkx==2.6.3
# nltk==3.7
# notebook==6.4.12
# numba==0.55.2
# numexpr==2.8.1
# numpy==1.19.5
# oauthlib==3.2.0
# packaging==21.3
# pandas==1.3.5
# pandas-profiling==3.2.0
# pandocfilters==1.5.0
# parso==0.8.3
# patsy==0.5.2
# phik==0.12.2
# pickleshare==0.7.5
# Pillow==9.1.1
# plac==1.1.3
# plotly==5.8.2
# preshed==3.0.6
# prometheus-client==0.14.1
# prometheus-flask-exporter==0.20.2
# prompt-toolkit==3.0.29
# protobuf==4.21.1
# psutil==5.9.1
# pycaret==2.3.0
# pycparser==2.21
# pydantic==1.9.1
# Pygments==2.12.0
# PyJWT==2.4.0
# pyLDAvis==3.2.2
# pynndescent==0.5.7
# pyod==1.0.1
# pyparsing==3.0.9
# pyrsistent==0.18.1
# python-dateutil==2.8.2
# pytz==2022.1
# PyWavelets==1.3.0
# # pywin32==227
# # pywinpty==2.0.5
# PyYAML==6.0
# pyzmq==23.1.0
# querystring-parser==1.2.4
# regex==2022.6.2
# requests==2.28.0
# requests-oauthlib==1.3.1
# scikit-learn==0.23.2
# scikit-plot==0.3.7
# scipy==1.5.4
# seaborn==0.11.2
# Send2Trash==1.8.0
# simplejson==3.17.6
# six==1.16.0
# smart-open==6.0.0
# smmap==5.0.0
# soupsieve==2.3.2.post1
# spacy==2.3.7
# SQLAlchemy==1.4.37
# sqlparse==0.4.2
# srsly==1.0.5
# statsmodels==0.13.2
# tabulate==0.8.9
# tangled-up-in-unicode==0.2.0
# tenacity==8.0.1
# terminado==0.15.0
# textblob==0.17.1
# thinc==7.4.5
# threadpoolctl==3.1.0
# tinycss2==1.1.1
# tornado==6.1
# tqdm==4.64.0
# traitlets==5.2.2.post1
# typing_extensions==4.2.0
# umap-learn==0.5.3
# urllib3==1.26.9
# visions==0.7.4
# waitress==2.1.2
# wasabi==0.9.1
# wcwidth==0.2.5
# webencodings==0.5.1
# websocket-client==1.3.2
# Werkzeug==2.1.2
# widgetsnbextension==3.6.0
# wordcloud==1.8.1
# xgboost==1.2.0
# yarl==1.7.2
# yellowbrick==1.3.post1
# zipp==3.8.0


aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
alembic==1.7.7
anyio==3.5.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
autopage==0.5.0
autopep8==1.6.0
azure-core==1.23.0
azure-storage-blob==12.10.0
azure-storage-file-datalake==12.6.0
bleach==4.1.0
blis==0.7.7
Boruta==0.3
botocore==1.24.25
cachetools==5.0.0
catalogue==1.0.0
catboost==1.0.4
category-encoders==2.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
cliff==3.10.1
cloudpickle==2.0.0
cmaes==0.8.2
cmd2==2.4.0
colorama==0.4.4
colorlog==6.6.0
colorlover==0.3.0
cryptography==36.0.2
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.6
Cython==0.29.23
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
docutils==0.15.2
ecdsa==0.17.0
entrypoints==0.4
et-xmlfile==1.1.0
filelock==3.6.0
fonttools==4.30.0
frozenlist==1.3.0
fsspec==2022.2.0
funcy==1.17
future==0.18.2
gensim==4.1.2
gitdb==4.0.9
GitPython==3.1.27
gower==0.0.5
graphviz==0.19.1
greenlet==1.1.2
grpcio==1.43.0
gunicorn==20.1.0
h11==0.13.0
htmlmin==0.1.12
httptools==0.4.0
hyperopt==0.2.7
idna==3.3
ImageHash==4.2.1
imbalanced-learn==0.9.0
importlib-metadata==4.11.3
importlib-resources==5.4.0
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.0
jmespath==1.0.0
joblib==1.0.1
jose==1.0.0
jsonschema==4.4.0
kiwisolver==1.4.0
kmodes==0.11.1
lightgbm==3.3.2
llvmlite==0.38.0
Mako==1.2.0
MarkupSafe==2.0.1
missingno==0.5.1
mistune==0.8.4
mlflow==1.24.0
mlxtend==0.19.0
msgpack==1.0.3
msrest==0.6.21
multidict==6.0.2
multimethod==1.7
murmurhash==1.0.6
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
networkx==2.6.3
nltk==3.7
notebook==6.4.10
numba==0.55.1
numexpr==2.8.1
numpy==1.19.5
oauthlib==3.2.0
openpyxl==3.0.10
optuna==2.10.0
packaging==21.3
pandas==1.1.4
pandas-profiling==3.1.0
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.2
pbr==5.8.1
phik==0.12.2
pickleshare==0.7.5
Pillow==9.0.1
plac==1.1.3
plotly==5.6.0
preshed==3.0.6
prettytable==3.2.0
prometheus-client==0.13.1
prometheus-flask-exporter==0.19.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
psycopg2==2.9.3
py4j==0.10.9.5
pyaml==21.10.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycaret==2.3.0
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.3.1
pydantic==1.9.0
Pygments==2.11.2
pyLDAvis==3.3.1
pynndescent==0.5.6
pyOpenSSL==22.0.0
pyparsing==3.0.7
pyperclip==1.8.2
pyreadline3==3.4.1
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-jose==3.3.0
python-jose-cryptodome==1.3.2
pytz==2021.3
PyWavelets==1.3.0
PyYAML==6.0
pyzmq==22.3.0
querystring-parser==1.2.4
ray==1.11.0
regex==2022.3.15
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.7.2
s3transfer==0.5.2
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.0
shap==0.40.0
simplejson==3.17.6
six==1.16.0
# sklearn==0.0
scikit-learn==0.23.2
slicer==0.0.7
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
soupsieve==2.3.1
spacy==2.3.7
SQLAlchemy==1.4.32
sqlparse==0.4.2
srsly==1.0.5
starlette==0.17.1
statsmodels==0.13.2
stevedore==3.5.0
tabulate==0.8.9
tangled-up-in-unicode==0.1.0
tenacity==8.0.1
tensorboardX==2.5
terminado==0.13.3
testpath==0.6.0
textblob==0.17.1
thinc==7.4.5
threadpoolctl==3.1.0
toml==0.10.2
tornado==6.1
tqdm==4.63.1
traitlets==5.1.1
tune-sklearn==0.4.1
typing_extensions==4.1.1
umap-learn==0.5.2
urllib3==1.23
uvicorn==0.17.6
visions==0.7.4
waitress==2.1.1
wasabi==0.9.0
watchgod==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.1
websockets==10.3
Werkzeug==2.0.3
widgetsnbextension==3.6.0
wordcloud==1.8.1
wrapt==1.14.0
xgboost==1.2.0
yarl==1.7.2
yellowbrick==1.4
zipp==3.7.0